requirements.txt
bcrypt==4.0.1
beautifultable==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
filelock==3.12.4
idna==3.4
paramiko==3.3.1
pycparser==2.21
PyNaCl==1.5.0
pysondb-v2==2.1.0
requests==2.31.0
urllib3==2.0.6
wcwidth==0.2.8
